+++ CHANGES.rst
@@ -5,4 +5,10 @@ Changelog of lizard-area
 0.1 (unreleased)
 ----------------
 
+- Created geoobject.py helper library.
+
+- Added admin.
+
+- Initial models.
+
 - Initial library skeleton created by nensskel.  [Jack Ha]